+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed: clarified error message on multiple and not string `userAgent` option.
 - Fixed: `Host` directive is now not containing `80` port.
 - Fixed: thrown error if `cleanParam` not string or array and if string not more than 500 characters.
+- Fixed: supported unicode characters in `Allow` and `Disallow` directives.
 
 # 4.0.1 - 2016-10-27